Fix synchronization bug
Fixed a bug where totalFishCaught is set to totalCratesCaught whenever the client updates the server/other clients with Koko Net Lib.

Since the packet relaying actually sends the packets back to the original sender, it causes the inaccuracy on the client who initially sent the packet as well :). Probs never gonna change this issue where the client "self-syncs" aka receives its own packet. I guess in this case it is a "self-desync" lol. Just too lazy to add an ignoreClient param to all of my my Koko  Lib code packets for the sake of changing something that doesn't rlly need to be changed.